chore: rename files and improving organization
